Weaponizer+Maskizer touch / better touch input / Intersector / Attacking state...
 - Weaponizer and Maskizer buttons not if tocuehd don't update player's
touchpos. With that came the next one.
 - Better Weaponizer and Maskizer collision detection: now we have our
vector2s intercepting with our circles.
  *FIX* : The fact that the Weaponizer can't get boolean doIntercept
from theController, thus having to create a vector2 of his own. Those
mistakes are marked by comments: "this is crap"
  - Enemy state ATTACKING has been modofied so that he can no longer
move, but is able to perform an attack waiting a certain time between
them. Also a bit fuzzy, try and refactor it a bit.
  
  **STABLE VERSION**