OxygenBar Flicker/ GameOverSceenUI FIXED/ WittyMessageGenerator/ ANIMATINGLARGE.State
- OxygenBar now flickers when the Oxygenbar is empty with a noticeable
colour.  
- GameOverUI was finally fixed, though now the first click will not do
anything, maybe change that later. 
- Introducing WittyMessageGenerator, in the GameOverScreenUI it now
randomly generates a string from the ones available in Constants, and
displays it in GameOverScreen 
- ANIMATIONLARGE may be a temporary state, because of it being slightly
impractical, but it will do for now, to get rid of it, you must find a
way to change the way the String Image is divided for a certain
animation without making a separate state for that.